feat: bump fiber-go to v1.9.3, configure health checks (#47)
## 📝 Summary

Bumps fiber-go to v1.9.3 and configures a health check. This health
check will detect when the connection is not in the `connectivity.Ready`
state and reconnect everything if that's the case. Attempts to fix the
silent streams issue.

> [!NOTE]
> Since we haven't been able to reproduce, this might still not fix it.
If that's the case we can try other methods of detection (like recording
when the last message was seen on active streams and using that to
detect the issue)

---

## ✅ I have run these commands

* [x] `make lint`
* [x] `make test`
* [x] `go mod tidy`