fix: make chainbound reconnecting more robust (#54)
## 📝 Summary

Simplify Chainbound reconnection logic.
<!--- A general summary of your changes -->

## ⛱ Motivation and Context

Before it had recursive start -> reconnect -> start, which was hard to
reason about and could lead to unbounded call-stack growth. Now there's
a single outer loop in `Start()`.

Instead of modifying the `ChainboundNodeConnection` across channels,
which is brittle, we now pass the channel as a function param to avoid
possible race condition.


---

## ✅ I have run these commands

* [x] `make lint`
* [x] `make test`
* [x] `go mod tidy`