Create transactions archive using Clickhouse as data source (#80)
## 📝 Summary

This PR enables creation of archives using Clickhouse as data source.

The transactions alone need to load already 10+ GB of data from
Clickhouse, and doesn't include the sources anymore. This is an
acceptable tradeoff, versus using a join on the data and potentially
3x/4x the amount of data needed to be loaded for adding the sources too.
A future PR might address this if this is a requested feature.

## ⛱ Motivation and Context

Mempool Dumpster added support for storing data in Clickhouse, which
allows for a better, redundant infrastructure setup (instead of relying
on local CSV files).